+++ main.go
@@ -29,12 +29,11 @@ type appConfig struct {
 	Debug          bool   `conf:"default:false,help:enable debug log level"`
 	Clipboard      struct {
 		MaxHistory   int           `conf:"default:50,help:maximum number of clipboard entries to keep in history"`
-		PollInterval time.Duration `conf:"default:500ms,help:how often to poll the system clipboard for changes"`
-		ForceXclip   bool          `conf:"default:false,help:force xclip backend instead of auto-detecting"`
+		PollInterval time.Duration `conf:"default:2s,help:in case we fallback to polling, how often to poll the system clipboard for changes"`
 	}
 	RemoteClipboards struct {
 		MaxHistory   int           `conf:"default:5,help:maximum number of local clipboard entries to transmit to remote peers"`
-		PollInterval time.Duration `conf:"default:1s,help:how often to fetch clipboard entries from remote peers"`
+		PollInterval time.Duration `conf:"default:2s,help:how often to fetch clipboard entries from remote peers"`
 		Disable      bool          `conf:"default:false,help:disable fetching clipboards from remote peers"`
 	}
 	Peers struct {
@@ -83,7 +82,6 @@ func run() error {
 		ThemeColorPath:               cfg.ThemeColorPath,
 		ConfigPath:                   cfg.ConfigPath,
 		PollInterval:                 cfg.Clipboard.PollInterval,
-		ForceXclip:                   cfg.Clipboard.ForceXclip,
 		RemoteClipboardsPollInterval: cfg.RemoteClipboards.PollInterval,
 		RemoteClipboardsMaxHistory:   cfg.RemoteClipboards.MaxHistory,
 		PeersPollInterval:            cfg.Peers.PollInterval,